+++ lib/govuk_schemas.rb
@@ -5,8 +5,14 @@ require "govuk_schemas/document_types"
 require "govuk_schemas/example"
 
 module GovukSchemas
-  # @private
-  CONTENT_SCHEMA_DIR = ENV["GOVUK_CONTENT_SCHEMAS_PATH"] || "../govuk-content-schemas"
+  def self.content_schema_dir=(content_schema_dir)
+    @content_schema_dir = content_schema_dir
+  end
+
+  def self.content_schema_dir
+    @content_schema_dir ||= ENV.fetch("GOVUK_CONTENT_SCHEMAS_PATH", "../govuk-content-schemas")
+    @content_schema_dir
+  end
 
   # @private
   class InvalidContentGenerated < RuntimeError